moved the css into a new styles folder for correct consistency of the file system
